split TreeContainer
